Add files via upload
This is a shiny app which can be use for explore a dataset. This is consisting with 7 panels; Data, Overview, One way, Two way, Three way & References panels.  In the first
two panels you get an idea about the dataset, what are the variables & overview of the data.  From the third panel, you can find which variables in the dataset are highly correlated or not. From the next three panels, you can see the exploratory analysis by one way, two way & three way respectively. If you need to know further details on R shiny you can access my references.